Iterative Fisher-Yates Exact test for collexeme/collocational analysis

This is a vectorised wrapper for the dhyper function in the stats package. The implementation of the code is adapted from Gries (2012). collex_fye also provides a logical argument (i.e., two_sided) whose value is passed to the alternative argument of the embedded fisher.test if two_sided is TRUE.

collex_fye(
  a,
  n_corpus = "total size of the corpus",
  n_coll = "total frequency of collexemes/collocates in the corpus as a whole",
  n_cxn = "total frequency of the construction/node word in the corpus as a whole",
  two_sided = FALSE,
  collstr_res = TRUE,
  float = 3,
  mpfr_precision = 500
)

Arguments

a
cell a in a 2-by-2 crosstabulation matrix (viz. representing the co-occurrence tokens of the levels of the variables. For instance, word-word co-occurrences, or word-construction co-occurrences).

n_corpus
the total size (in word tokens) of the corpus.

n_coll
the total frequency of the collexemes/collocates of the target construction/node word in the corpus.

n_cxn
the total frequency of occurrence of the target construction/node word in the corpus.

two_sided
logical; whether to perform one-sided test (FALSE -- Default) or two-sided (TRUE).

collstr_res
logical; whether output the FYE p-value as the Collostruction Strength value (TRUE -- the default) or just report the p-value (FALSE).

float
the floating digits of the Collostruction Strength. The default is 3.

mpfr_precision
a number indicating the maximal precision to be used in bits. This is passed to the precBits argument of mpfr.

Value

Numeric vector of the same length as a interpreted as the Collostruction Strength of the construction/node word with the collexemes/collocates. Collostruction Strength is (i) the negative logarithm to the base of ten of the Fisher-Yates Exact test p-value when a > a_exp, and (ii) the positive logarithm when a <= a_exp.