Read in the wordlist files of the Leipzig Corpora

The function generates a tibble of frequency list of words read from the Leipzig Corpora frequency list files. When one downloads the zip file of a Leipzig corpus, the readily available frequency lists will have the "-words.txt" label. For instance, for Indonesian, the filename looks like this: "ind_newscrawl_2015_300K-words.txt".

freqlist_leipzig_read(
  leipzig_wlist_path = leipzig_wordlist_path,
  file_pattern = NULL,
  all = FALSE,
  lower_case = TRUE
)

Arguments

leipzig_wlist_path
the file paths of the Leipzig wordlist files.

file_pattern
the regular expressions for the wordlist files if particular files are to be read. The default is NULL. When all available frequency list files want to be be read-in, leave this argument to NULL and set the all argument (cf. below) to TRUE.

all
whether to read all available wordlist files from all corpus (TRUE) or not (FALSE -- the default)

lower_case
whether the words are lowercased (TRUE -- the default) or not (FALSE)

Value

a tibble with three columns (i) corpus_id (i.e., the corpus file), (ii) word (i.e., word forms), and (iii) n (i.e., token frequency)

Examples

if (FALSE) {
# read in a file containing filepaths of all Leipzig Corpora frequency list files
wordlist_path <- readr::read_lines(file = "~/Documents/Corpora/leipzig_wordlist_path.txt")

# read in the web and wikipedia sections
tb <- freqlist_leipzig_read(leipzig_wordlist_path,
                            file_pattern = "(web_2012|wikipedia_2016)",
                            all = FALSE,
                            lower_case = TRUE)

 # subset only on word that consists of letters and hypen
 # to retrieve/maintain reduplication (e.g., "anak-anak" 'children')
 dplyr::filter(tb, stringr::str_detect(word, "^([-a-z]+)$"))
}